Refactor builtins to use Builtin trait
- Define Builtin trait with name(), description(), and run() methods
- Convert all builtin commands to implement the trait as structs
- Update execute_builtin to dynamically dispatch using trait objects
- Modify help command to retrieve builtin info from trait implementations
- Maintain all existing functionality and tests pass
