+++ src/builtins.rs
@@ -1,53 +1,55 @@
-use std::env;
 use std::fs::File;
 use std::io::{self, Write};
-use std::path::Path;
 
 use crate::parser::ShellCommand;
 use crate::state::ShellState;
 
-const BUILTINS: &[&str] = &[
-    "cd", "echo", "pwd", "env", "exit", "help", "source", "export", "unset", "pushd", "popd",
-    "dirs", "alias", "unalias",
-];
+mod builtin_alias;
+mod builtin_cd;
+mod builtin_dirs;
+mod builtin_echo;
+mod builtin_env;
+mod builtin_exit;
+mod builtin_export;
+mod builtin_help;
+mod builtin_popd;
+mod builtin_pushd;
+mod builtin_pwd;
+mod builtin_source;
+mod builtin_unalias;
+mod builtin_unset;
 
-const BUILTIN_DESCRIPTIONS: &[(&str, &str)] = &[
-    ("cd", "Change directory"),
-    ("echo", "Print arguments"),
-    ("pwd", "Print working directory"),
-    ("env", "Print environment variables"),
-    ("exit", "Exit the shell"),
-    ("help", "Show this help message"),
-    ("source", "Execute a script file with rush"),
-    ("export", "Export variables to environment"),
-    ("unset", "Unset shell variables"),
-    ("pushd", "Push directory onto stack and change to it"),
-    ("popd", "Pop directory from stack and change to it"),
-    ("dirs", "Display directory stack"),
-    ("alias", "Define or display aliases"),
-    ("unalias", "Remove alias definitions"),
-];
+pub trait Builtin {
+    fn name(&self) -> &'static str;
+    fn description(&self) -> &'static str;
+    fn run(&self, cmd: &ShellCommand, shell_state: &mut ShellState, output_writer: &mut dyn Write) -> i32;
+}
+
+fn get_builtins() -> Vec<Box<dyn Builtin>> {
+    vec![
+        Box::new(builtin_cd::CdBuiltin),
+        Box::new(builtin_echo::EchoBuiltin),
+        Box::new(builtin_pwd::PwdBuiltin),
+        Box::new(builtin_env::EnvBuiltin),
+        Box::new(builtin_exit::ExitBuiltin),
+        Box::new(builtin_help::HelpBuiltin),
+        Box::new(builtin_source::SourceBuiltin),
+        Box::new(builtin_export::ExportBuiltin),
+        Box::new(builtin_unset::UnsetBuiltin),
+        Box::new(builtin_pushd::PushdBuiltin),
+        Box::new(builtin_popd::PopdBuiltin),
+        Box::new(builtin_dirs::DirsBuiltin),
+        Box::new(builtin_alias::AliasBuiltin),
+        Box::new(builtin_unalias::UnaliasBuiltin),
+    ]
+}
 
 pub fn is_builtin(cmd: &str) -> bool {
-    BUILTINS.contains(&cmd)
+    get_builtins().iter().any(|b| b.name() == cmd)
 }
 
 pub fn get_builtin_commands() -> Vec<String> {
-    BUILTINS.iter().map(|s| s.to_string()).collect()
-}
-
-fn print_dir_stack<W: Write>(shell_state: &ShellState, writer: &mut W) {
-    // Get current directory
-    if let Ok(current) = env::current_dir() {
-        let current_str = current.to_string_lossy().to_string();
-        // Print current directory first
-        let _ = write!(writer, "{}", current_str);
-        // Then print stack in reverse order (top of stack first)
-        for dir in shell_state.dir_stack.iter().rev() {
-            let _ = write!(writer, " {}", dir);
-        }
-        let _ = writeln!(writer);
-    }
+    get_builtins().iter().map(|b| b.name().to_string()).collect()
 }
 
 pub fn execute_builtin(
@@ -91,275 +93,11 @@ pub fn execute_builtin(
         Box::new(io::stdout())
     };
 
-    match cmd.args[0].as_str() {
-        "cd" => {
-            // cd doesn't produce output, so redirections don't make sense for it
-            // But we still handle it for consistency
-            let dir = if cmd.args.len() > 1 {
-                cmd.args[1].clone()
-            } else {
-                "~".to_string()
-            };
-            let path = if dir == "~" {
-                env::var("HOME").unwrap_or_else(|_| "/".to_string())
-            } else {
-                dir
-            };
-            if let Err(e) = env::set_current_dir(Path::new(&path)) {
-                let _ = writeln!(output_writer, "cd: {}: {}", path, e);
-                1
-            } else {
-                0
-            }
-        }
-        "echo" => {
-            let output = cmd.args[1..].join(" ");
-            let _ = writeln!(output_writer, "{}", output);
-            0
-        }
-        "pwd" => match env::current_dir() {
-            Ok(path) => {
-                let _ = writeln!(output_writer, "{}", path.display());
-                0
-            }
-            Err(e) => {
-                let _ = writeln!(output_writer, "pwd: {}", e);
-                1
-            }
-        },
-        "env" => {
-            // Show exported shell variables first
-            for var_name in &shell_state.exported {
-                if let Some(value) = shell_state.variables.get(var_name) {
-                    let _ = writeln!(output_writer, "{}={}", var_name, value);
-                }
-            }
-            // Then show environment variables (excluding those already shown)
-            for (key, value) in env::vars() {
-                if !shell_state.exported.contains(&key) {
-                    let _ = writeln!(output_writer, "{}={}", key, value);
-                }
-            }
-            0
-        }
-        "exit" => {
-            // For now, just return 0; main handles exit
-            0
-        }
-        "help" => {
-            // Attempt to write the header, handling potential errors
-            if writeln!(output_writer, "Rush Shell v{}", env!("CARGO_PKG_VERSION")).is_err()
-                || writeln!(output_writer, "").is_err()
-                || writeln!(output_writer, "Available built-in commands:").is_err()
-            {
-                return 1; // Return error if header write fails
-            }
-
-            // Iterate over the complete builtins list with descriptions
-            for (cmd, desc) in BUILTIN_DESCRIPTIONS {
-                // Use explicit formatting for better readability
-                let formatted_line = format!("  {:<12} {}", cmd, desc);
-                if writeln!(output_writer, "{}", formatted_line).is_err() {
-                    return 1; // Return error if any command write fails
-                }
-            }
-            0
-        }
-        "source" => {
-            if cmd.args.len() < 2 {
-                let _ = writeln!(output_writer, "source: missing script file operand");
-                return 1;
-            }
-            let script_file = &cmd.args[1];
-
-            match std::fs::read_to_string(script_file) {
-                Ok(content) => {
-                    let mut exit_code = 0;
-                    for line in content.lines() {
-                        let line = line.trim();
-                        // Skip shebang lines and empty lines
-                        if line.is_empty() || line.starts_with("#!") {
-                            continue;
-                        }
-                        // Skip comment lines
-                        if line.starts_with("#") {
-                            continue;
-                        }
-                        // Execute the line using the same logic as main.rs
-                        // Note: source builtin doesn't have access to shell state, so we create a temporary one
-                        // This is a limitation - sourced scripts don't share variables with parent
-                        let mut temp_state = crate::state::ShellState::new();
-                        match crate::lexer::lex(line, &temp_state) {
-                            Ok(tokens) => match crate::parser::parse(tokens) {
-                                Ok(ast) => {
-                                    exit_code = crate::executor::execute(ast, &mut temp_state);
-                                }
-                                Err(e) => {
-                                    let _ = writeln!(output_writer, "Parse error: {}", e);
-                                    return 1;
-                                }
-                            },
-                            Err(e) => {
-                                let _ = writeln!(output_writer, "Lex error: {}", e);
-                                return 1;
-                            }
-                        }
-                    }
-                    exit_code
-                }
-                Err(e) => {
-                    let _ = writeln!(output_writer, "source: {}: {}", script_file, e);
-                    1
-                }
-            }
-        }
-        "export" => {
-            if cmd.args.len() < 2 {
-                // Print all exported variables
-                for var_name in &shell_state.exported {
-                    if let Some(value) = shell_state.variables.get(var_name) {
-                        let _ = writeln!(output_writer, "export {}={}", var_name, value);
-                    }
-                }
-                0
-            } else {
-                let arg = &cmd.args[1];
-                if let Some(eq_pos) = arg.find('=') {
-                    // export VAR=value
-                    let var = arg[..eq_pos].to_string();
-                    let value = arg[eq_pos + 1..].to_string();
-                    shell_state.set_exported_var(&var, value);
-                    0
-                } else {
-                    // export VAR (mark existing var as exported)
-                    shell_state.export_var(arg);
-                    0
-                }
-            }
-        }
-        "unset" => {
-            if cmd.args.len() < 2 {
-                let _ = writeln!(output_writer, "unset: missing variable name");
-                1
-            } else {
-                shell_state.unset_var(&cmd.args[1]);
-                0
-            }
-        }
-        "pushd" => {
-            if cmd.args.len() < 2 {
-                let _ = writeln!(output_writer, "pushd: missing directory operand");
-                1
-            } else {
-                let dir = &cmd.args[1];
-                let path = if dir == "~" {
-                    env::var("HOME").unwrap_or_else(|_| "/".to_string())
-                } else {
-                    dir.clone()
-                };
-
-                // Get current directory before changing
-                let current_dir = match env::current_dir() {
-                    Ok(path) => path.to_string_lossy().to_string(),
-                    Err(e) => {
-                        let _ = writeln!(
-                            output_writer,
-                            "pushd: failed to get current directory: {}",
-                            e
-                        );
-                        return 1;
-                    }
-                };
-
-                // Change to new directory
-                if let Err(e) = env::set_current_dir(Path::new(&path)) {
-                    let _ = writeln!(output_writer, "pushd: {}: {}", path, e);
-                    1
-                } else {
-                    // Push old directory to stack
-                    shell_state.dir_stack.push(current_dir);
-                    // Print the stack
-                    print_dir_stack(shell_state, &mut output_writer);
-                    0
-                }
-            }
-        }
-        "popd" => {
-            if shell_state.dir_stack.is_empty() {
-                let _ = writeln!(output_writer, "popd: directory stack empty");
-                1
-            } else {
-                // Pop directory from stack
-                let dir = shell_state.dir_stack.pop().unwrap();
-                // Change to that directory
-                if let Err(e) = env::set_current_dir(Path::new(&dir)) {
-                    let _ = writeln!(output_writer, "popd: {}: {}", dir, e);
-                    1
-                } else {
-                    // Print the stack
-                    print_dir_stack(shell_state, &mut output_writer);
-                    0
-                }
-            }
-        }
-        "dirs" => {
-            print_dir_stack(shell_state, &mut output_writer);
-            0
-        }
-        "alias" => {
-            if cmd.args.len() == 1 {
-                // List all aliases
-                let aliases = shell_state.get_all_aliases();
-                if aliases.is_empty() {
-                    let _ = writeln!(output_writer);
-                } else {
-                    for (name, value) in aliases {
-                        let _ = writeln!(output_writer, "alias {}='{}'", name, value);
-                    }
-                }
-                0
-            } else if cmd.args.len() == 2 {
-                let arg = &cmd.args[1];
-                if let Some(eq_pos) = arg.find('=') {
-                    // Set alias: alias name=value
-                    let name = arg[..eq_pos].to_string();
-                    let value = arg[eq_pos + 1..].to_string();
-                    shell_state.set_alias(&name, value);
-                    0
-                } else {
-                    // Show specific alias
-                    if let Some(value) = shell_state.get_alias(arg) {
-                        let _ = writeln!(output_writer, "alias {}='{}'", arg, value);
-                        0
-                    } else {
-                        let _ = writeln!(output_writer, "alias: {}: not found", arg);
-                        1
-                    }
-                }
-            } else {
-                let _ = writeln!(output_writer, "alias: too many arguments");
-                1
-            }
-        }
-        "unalias" => {
-            if cmd.args.len() < 2 {
-                let _ = writeln!(output_writer, "unalias: missing alias name");
-                1
-            } else if cmd.args.len() > 2 {
-                let _ = writeln!(output_writer, "unalias: too many arguments");
-                1
-            } else {
-                let name = &cmd.args[1];
-                if shell_state.get_alias(name).is_some() {
-                    shell_state.remove_alias(name);
-                    0
-                } else {
-                    let _ = writeln!(output_writer, "unalias: {}: not found", name);
-                    1
-                }
-            }
-        }
-        _ => 1,
+    let builtins = get_builtins();
+    if let Some(builtin) = builtins.into_iter().find(|b| b.name() == cmd.args[0]) {
+        builtin.run(cmd, shell_state, &mut *output_writer)
+    } else {
+        1
     }
 }
